requirements.txt
Flask==3.0.0
line-bot-sdk==3.21.0
gunicorn==21.2.0